requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
assertpy==1.1
attrs==23.1.0
black==24.3.0
certifi==2024.07.04
charset-normalizer==3.1.0
curlify==2.2.1
enum34==1.1.10
flake8==5.0.4
idna==3.7
iniconfig==2.0.0
lxml==5.3.0
namedlist==1.8
packaging==23.1
pluggy==1.0.0
pylint==2.17.4
pytest==7.3.1
pytest-retry==1.3
python-dateutil==2.8.2
requests==2.32.0
six==1.16.0
urllib3==1.26.19